// src/lib/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      socio_titulares: {
        Row: {
          id: string
          created_at: string
          dni: string
          nombres: string
          apellidoPaterno: string
          apellidoMaterno: string
          celular: string | null
          localidad: string
          mz: string | null
          lote: string | null
          is_payment_observed: boolean
          payment_observation_detail: string | null
          is_lote_medido: boolean
          isObservado: boolean
          observacion: string | null
        }
        Insert: {
          id?: string
          created_at?: string
          dni: string
          nombres: string
          apellidoPaterno: string
          apellidoMaterno: string
          celular?: string | null
          localidad: string
          mz?: string | null
          lote?: string | null
          is_payment_observed?: boolean
          payment_observation_detail?: string | null
          is_lote_medido?: boolean
          isObservado?: boolean
          observacion?: string | null
        }
        Update: {
          id?: string
          created_at?: string
          dni?: string
          nombres?: string
          apellidoPaterno?: string
          apellidoMaterno?: string
          celular?: string | null
          localidad?: string
          mz?: string | null
          lote?: string | null
          is_payment_observed?: boolean
          payment_observation_detail?: string | null
          is_lote_medido?: boolean
          isObservado?: boolean
          observacion?: string | null
        }
      }
      ingresos: {
        Row: {
          id: number
          created_at: string
          receipt_number: string
          dni: string | null
          full_name: string | null
          amount: number
          account: string
          date: string
          transaction_type: string | null
          numeroOperacion: number | null
        }
        Insert: {
          id?: number
          created_at?: string
          receipt_number: string
          dni?: string | null
          full_name?: string | null
          amount: number
          account: string
          date: string
          transaction_type?: string | null
          numeroOperacion?: number | null
        }
        Update: {
          id?: number
          created_at?: string
          receipt_number?: string
          dni?: string | null
          full_name?: string | null
          amount?: number
          account?: string
          date?: string
          transaction_type?: string | null
          numeroOperacion?: number | null
        }
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
  }
}

export type Tables<T extends keyof Database['public']['Tables']> = Database['public']['Tables'][T]['Row']
export type TablesInsert<T extends keyof Database['public']['Tables']> = Database['public']['Tables'][T]['Insert']